Add tests for metadata tmp dir
Requires refactoring existing tests and migrating to Junit 5
