+++ hadoop-hdds/container-service/src/main/java/org/apache/hadoop/ozone/container/common/volume/MetadataVolume.java
@@ -33,11 +33,23 @@ public class MetadataVolume extends StorageVolume {
     super(b);
     // Tmp directory on Metadata Volume uses the volume root as the working
     // directory. It is not dependent on the cluster ID from SCM.
-    createTmpDirs("");
+    super.createTmpDirs("");
   }
 
   public VolumeType getType() {
     return type;
+  }
+
+  @Override
+  public void format(String cid) throws IOException {
+    // No-op for Metadata volumes.
+  }
+
+  @Override
+  public void createTmpDirs(String workingDirName) {
+    // No-op for metadata volumes.
+    // Tmp directory is created on construction since it is not dependent on
+    // getting the cluster ID from SCM.
   }
 
   /**
